Tidied up Anamnesis source code
